backend/requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework>=3.12.4,<3.13
django-cors-headers
pymysql
djangorestframework
djangorestframework-simplejwt
Pillow
pytz
cryptography
channels>=3.0.5
daphne
channels_redis